+++ packages/demo/vite.config.ts
@@ -1,5 +1,5 @@
-import { globRoutes } from "@hiogawa/vite-glob-routes";
-import { indexHtmlMiddlewarePlugin } from "@hiogawa/vite-index-html-middleware";
+import globRoutesPlugin from "@hiogawa/vite-glob-routes";
+import indexHtmlMiddlewarePlugin from "@hiogawa/vite-index-html-middleware";
 import vaviteConnect from "@vavite/connect";
 import react from "@vitejs/plugin-react";
 import unocss from "unocss/vite";
@@ -11,7 +11,7 @@ export default defineConfig((ctx) => ({
   plugins: [
     unocss(),
     react(),
-    globRoutes({ root: "/src/routes" }),
+    globRoutesPlugin({ root: "/src/routes" }),
     indexHtmlMiddlewarePlugin(),
     vaviteConnect({
       standalone: false,
