+++ packages/vite-glob-routes/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@hiogawa/vite-glob-routes",
-  "version": "0.1.0",
+  "version": "0.1.1",
   "type": "module",
   "main": "./dist/index.js",
   "module": "./dist/index.js",
@@ -11,17 +11,17 @@
       "require": "./dist/index.cjs",
       "types": "./dist/index.d.ts"
     },
-    "./internal": {
+    "./dist/internal": {
       "import": "./dist/internal/index.js",
       "require": "./dist/internal/index.cjs",
       "types": "./dist/internal/index.d.ts"
     },
-    "./react-router": {
+    "./dist/react-router": {
       "import": "./dist/react-router.js",
       "require": "./dist/react-router.cjs",
       "types": "./dist/react-router.d.ts"
     },
-    "./hattip": {
+    "./dist/hattip": {
       "import": "./dist/hattip.js",
       "require": "./dist/hattip.cjs",
       "types": "./dist/hattip.d.ts"
