+++ packages/vite-glob-routes/src/index.ts
@@ -1,16 +1,17 @@
 import type { Plugin } from "vite";
 
-const PAGE_ROUTES_REACT_ROUTER = "@hiogawa/vite-glob-routes/react-router";
-const API_ROUTES_HATTIP = "@hiogawa/vite-glob-routes/hattip";
+const PAGE_ROUTES_REACT_ROUTER = "@hiogawa/vite-glob-routes/dist/react-router";
+const API_ROUTES_HATTIP = "@hiogawa/vite-glob-routes/dist/hattip";
 const TAG = "virtual:";
 
-export function globRoutes(options: { root: string }): Plugin[] {
+export default function globRoutesPlugin(options: { root: string }): Plugin[] {
   // TODO: escape js string
   const root = options.root;
   return [
     {
       name: "@hiogawa/vite-glob-routes",
-      enforce: "pre",
+      // https://github.com/vitejs/vite/issues/9364#issuecomment-1197842127
+      enforce: "pre", // TODO: this works in the demo (inside monorepo), but doesn't when installed externally in https://github.com/hi-ogawa/youtube-dl-web-v2/pull/62 ???
 
       async resolveId(source, _importer, _options) {
         if (
@@ -29,18 +30,17 @@ export function globRoutes(options: { root: string }): Plugin[] {
 
         if (id === TAG + PAGE_ROUTES_REACT_ROUTER) {
           return `
-            import { createGlobPageRoutes } from "@hiogawa/vite-glob-routes/internal";
+            import { createGlobPageRoutes } from "@hiogawa/vite-glob-routes/dist/internal";
             // TODO: non-eager import to code split?
             const globPage = import.meta.glob("${root}/**/*.page.(js|jsx|ts|tsx)", { eager: true });
             const globLayout = import.meta.glob("${root}/**/layout.(js|jsx|ts|tsx)", { eager: true });
             export const globPageRoutes = () => createGlobPageRoutes("${root}", globPage, globLayout);
-
           `;
         }
 
         if (id === TAG + API_ROUTES_HATTIP) {
           return `
-            import { createGlobApiRoutes } from "@hiogawa/vite-glob-routes/internal";
+            import { createGlobApiRoutes } from "@hiogawa/vite-glob-routes/dist/internal";
             // TODO: import only "get/post/put/delete" to tree shake other exports?
             const globApi = import.meta.glob("${root}/**/*.api.(js|jsx|ts|tsx)", { eager: true });
             export const globApiRoutes = () => createGlobApiRoutes("${root}", globApi);
