+++ packages/vite-index-html-middleware/src/internal.ts
